+++ core/include/booker/archive/RarWriter.hpp
@@ -1,16 +1,23 @@
 #ifndef DEF_RARWRITER
 #define DEF_RARWRITER
 
-/*
-Just a simple class to expose these methods :
-
-bool isRarInstalled() const;
-bool compressOneFile(std::string const& filePath, std::string const& archiveName, bool checkUnsafeChars = true) const;
-bool compressMultipleFiles(std::vector<std::string> const& files, std::string const& archiveName) const;
-bool compressDirectory(std::string const& directoryPath, std::string const& archiveName, bool checkUnsafeChars = true) const;
+/*!
+\file RarWriter.hpp
+\brief Define RarWriter.
+\details This file define RarWriter, a basic class to manage RAR writing.
 */
 
 #include <RarLib/Rar.hpp>
+
+/*!
+\class RarWriter RarWriter.hpp "core/include/booker/archive/RarWriter.hpp"
+\brief A class allowing to compress RAR.
+\details This class, by heriting from Rar (from RarLib), exposes these methods :
+- bool isRarInstalled() const;
+- bool compressOneFile(std::string const& filePath, std::string const& archiveName, bool checkUnsafeChars = true) const;
+- bool compressMultipleFiles(std::vector<std::string> const& files, std::string const& archiveName) const;
+- bool compressDirectory(std::string const& directoryPath, std::string const& archiveName, bool checkUnsafeChars = true) const;
+*/
 
 class RarWriter : public Rar
 {