+++ core/include/booker/archive/ZipException.hpp
@@ -1,7 +1,18 @@
 #ifndef DEF_ZIPEXCEPTION
 #define DEF_ZIPEXCEPTION
 
+/*!
+\file ZipException.hpp
+\brief Define ZipException.
+\details This file define ZipException, a class herited from \ref Exception class.
+*/
+
 #include "../utils/Exception.hpp"
+
+/*!
+\class ZipException ZipException.hpp "core/include/booker/archive/ZipException.hpp"
+\brief An Exception derived class, which indicates file, line and function name of the thrown exception.
+*/
 
 class ZipException : public Exception
 {