+++ core/include/booker/io/DirectoryUtils.hpp
@@ -1,8 +1,17 @@
 #ifndef DIRECTORYUTILS_HPP
 #define DIRECTORYUTILS_HPP
 
+/*!
+\file DirectoryUtils.hpp
+\brief Define utils functions related to directories.
+*/
+
 #include <filesystem>
 
+/*!
+\fn bool flattenDirectory(std::filesystem::path const& dirPath)
+\brief A function which make a directory flat.
+*/
 bool flattenDirectory(std::filesystem::path const& dirPath);
 
 #endif // DIRECTORYUTILS_HPP