+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - DirectoryUtils files to get utilities functions
 - Document sub-module in core, to handle very basic and primitive ebook
 - Format sub-module in core, to handle very basic and primitive ebook
+- Set up Doxygen documentation
 
 ### Changed