+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.11] - 2026-05-18
+
 ### Added
 
 - Pull Request build workflow
@@ -14,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improving README
 - Re-created sample CBZ and CBR, there was a format error for images before
 - Replacing unarr by libarchive
+- Conversions between CBR, CBZ and PDF are OK at this point
 
 ### Changed
 
@@ -123,7 +126,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Basic initialisation of repository
 - Basic function, converting each file of a PDF into JPG, to test GitHub CI configuration - this function will be removed
 
-[unreleased]: https://github.com/DostLeFan/Booker/compare/v0.0.10...HEAD
+[unreleased]: https://github.com/DostLeFan/Booker/compare/v0.0.11...HEAD
+[0.0.11]: https://github.com/DostLeFan/Booker/compare/v0.0.10...v0.0.11
 [0.0.10]: https://github.com/DostLeFan/Booker/compare/v0.0.9...v0.0.10
 [0.0.9]: https://github.com/DostLeFan/Booker/compare/v0.0.8...v0.0.9
 [0.0.8]: https://github.com/DostLeFan/Booker/compare/v0.0.7...v0.0.8